Reuse existing solutions for ErrorPage options
